feat: createListingComposable factory
