+++ packages/composables/src/index.ts
@@ -19,4 +19,5 @@ export * from "./logic/useNotifications";
 export * from "./logic/useIntercept";
 export * from "./getDefaultApiParams";
 export * from "./logic/useListing";
+export * from "./factories/createListingComposable";
 export { getApplicationContext, ApplicationVueContext } from "./appContext";
